fix: use proper endpoints. move changes to proper endpoint files, fix imports
